Add reward gold when killing enemies
- Each prefab enemies comes with two new script : Death and animation
- EnemyDeath : Manage death enemy (disable collider, time of destroy,etc)
- EnemyAnimation : if there is an animation -> Destroy needs to be triggered with a delay
- Preparation for Death animation on enemy, if the enemy attack the nuclear core or if he moves (Death_Vehicule and Death_Attack)

+++ Assets/Scripts/Enemies/EnemyStats.cs
@@ -5,7 +5,8 @@ public class EnemyStats : MonoBehaviour
     [Header("Stats")]
     public float EnemyHealth;
     public float EnemySpeed;
-    public float EnemyWorth;
+    //public float EnemyWorth; 
+    public int EnemyWorth;
     public float EnemyResistance;
     public float EnemyRange;
     public float EnemyFireRate;
